refactor: Extract github platform resolver
